:recycle: Modularize motion correction
