+++ CPAC/func_preproc/func_motion.py
@@ -1,0 +1,340 @@
+# Copyright (C) 2023  C-PAC Developers
+
+# This file is part of C-PAC.
+
+# C-PAC is free software: you can redistribute it and/or modify it under
+# the terms of the GNU Lesser General Public License as published by the
+# Free Software Foundation, either version 3 of the License, or (at your
+# option) any later version.
+
+# C-PAC is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
+# License for more details.
+
+# You should have received a copy of the GNU Lesser General Public
+# License along with C-PAC. If not, see <https://www.gnu.org/licenses/>.
+"""Functions for calculating motion parameters"""
+from nipype.interfaces.afni import preprocess
+from nipype.interfaces import fsl, utility as util
+from nipype.interfaces.afni import utils as afni_utils
+from CPAC.func_preproc.utils import chunk_ts, oned_text_concat, \
+                                    split_ts_chunks
+from CPAC.pipeline import nipype_pipeline_engine as pe
+from CPAC.pipeline.schema import latest_schema
+from CPAC.utils.interfaces.function import Function
+try:
+    motion_correct_options = latest_schema.schema['functional_preproc'][
+        'motion_estimates_and_correction'
+        ]['motion_correction']['using'][0].container
+except (AttributeError, KeyError, LookupError, IndexError):
+    motion_correct_options = ['3dvolreg', 'mcflirt']
+
+
+def get_mcflirt_rms_abs(rms_files):
+    for path in rms_files:
+        if 'abs.rms' in path:
+            abs_file = path
+        if 'rel.rms' in path:
+            rels_file = path
+    return abs_file, rels_file
+
+
+def motion_correct_3dvolreg(wf, cfg, strat_pool, pipe_num):
+    """Calculate motion parameters with 3dvolreg"""
+    if int(cfg.pipeline_setup['system_config']['max_cores_per_participant']
+           ) > 1:
+        chunk_imports = ['import nibabel as nb']
+        chunk = pe.Node(Function(input_names=['func_file',
+                                              'n_chunks',
+                                              'chunk_size'],
+                                  output_names=['TR_ranges'],
+                                  function=chunk_ts,
+                                  imports=chunk_imports),
+                        name=f'chunk_{pipe_num}')
+
+        #chunk.inputs.n_chunks = int(cfg.pipeline_setup['system_config'][
+        #                              'max_cores_per_participant'])
+
+        # 10-TR sized chunks
+        chunk.inputs.chunk_size = 10
+
+        node, out = strat_pool.get_data("desc-preproc_bold")
+        wf.connect(node, out, chunk, 'func_file')
+
+        split_imports = ['import os', 'import subprocess']
+        split = pe.Node(Function(input_names=['func_file',
+                                              'tr_ranges'],
+                                  output_names=['split_funcs'],
+                                  function=split_ts_chunks,
+                                  imports=split_imports),
+                        name=f'split_{pipe_num}')
+
+        node, out = strat_pool.get_data("desc-preproc_bold")
+        wf.connect(node, out, split, 'func_file')
+        wf.connect(chunk, 'TR_ranges', split, 'tr_ranges')
+
+        out_split_func = pe.Node(
+            interface=util.IdentityInterface(fields=['out_file']),
+            name=f'out_split_func_{pipe_num}')
+
+        wf.connect(split, 'split_funcs', out_split_func, 'out_file')
+
+        func_motion_correct = pe.MapNode(interface=preprocess.Volreg(),
+                                          name='func_generate_'
+                                              f'ref_{pipe_num}',
+                                          iterfield=['in_file'])
+
+        wf.connect(out_split_func, 'out_file',
+                    func_motion_correct, 'in_file')
+
+        func_concat = pe.Node(interface=afni_utils.TCat(),
+                              name=f'func_concat_{pipe_num}')
+        func_concat.inputs.outputtype = 'NIFTI_GZ'
+
+        wf.connect(func_motion_correct, 'out_file',
+                    func_concat, 'in_files')
+
+        out_motion = pe.Node(
+            interface=util.IdentityInterface(fields=['out_file']),
+            name=f'out_motion_{pipe_num}')
+
+        wf.connect(func_concat, 'out_file', out_motion, 'out_file')
+
+    else:
+        out_split_func = pe.Node(
+            interface=util.IdentityInterface(fields=['out_file']),
+            name=f'out_split_func_{pipe_num}')
+
+        node, out = strat_pool.get_data('desc-preproc_bold')
+        wf.connect(node, out, out_split_func, 'out_file')
+
+        func_motion_correct = pe.Node(interface=preprocess.Volreg(),
+                                      name=f'func_generate_ref_{pipe_num}')
+
+        wf.connect(out_split_func, 'out_file',
+                    func_motion_correct, 'in_file')
+
+        out_motion = pe.Node(
+            interface=util.IdentityInterface(fields=['out_file']),
+            name=f'out_motion_{pipe_num}')
+
+        wf.connect(func_motion_correct, 'out_file',
+                    out_motion, 'out_file')
+
+    func_motion_correct.inputs.zpad = 4
+    func_motion_correct.inputs.outputtype = 'NIFTI_GZ'
+
+    args = '-Fourier'
+    if cfg.functional_preproc['motion_estimates_and_correction'][
+        'motion_correction']['AFNI-3dvolreg']['functional_volreg_twopass']:
+        args = f'-twopass {args}'
+
+    func_motion_correct.inputs.args = args
+
+    # Calculate motion parameters
+    func_motion_correct_A = func_motion_correct.clone(
+        f'func_motion_correct_3dvolreg_{pipe_num}')
+    func_motion_correct_A.inputs.md1d_file = 'max_displacement.1D'
+    func_motion_correct_A.inputs.args = args
+
+    wf.connect(out_split_func, 'out_file',
+                func_motion_correct_A, 'in_file')
+
+    node, out = strat_pool.get_data('motion-basefile')
+    wf.connect(node, out, func_motion_correct_A, 'basefile')
+
+    if int(cfg.pipeline_setup['system_config'][
+                'max_cores_per_participant']) > 1:
+        motion_concat = pe.Node(interface=afni_utils.TCat(),
+                                name=f'motion_concat_{pipe_num}')
+        motion_concat.inputs.outputtype = 'NIFTI_GZ'
+
+        wf.connect(func_motion_correct_A, 'out_file',
+                    motion_concat, 'in_files')
+
+        out_motion_A = pe.Node(
+            interface=util.IdentityInterface(fields=['out_file']),
+            name=f'out_motion_A_{pipe_num}')
+
+        wf.connect(motion_concat, 'out_file',
+                    out_motion_A, 'out_file')
+
+        concat_imports = ['import os']
+        md1d_concat = pe.Node(Function(input_names=['in_files'],
+                                        output_names=['out_file'],
+                                        function=oned_text_concat,
+                                        imports=concat_imports),
+                                name=f'md1d_concat_{pipe_num}')
+
+        wf.connect(func_motion_correct_A, 'md1d_file',
+                    md1d_concat, 'in_files')
+
+        oned_concat = pe.Node(Function(input_names=['in_files'],
+                                        output_names=['out_file'],
+                                        function=oned_text_concat,
+                                        imports=concat_imports),
+                                name=f'oned_concat_{pipe_num}')
+
+        wf.connect(func_motion_correct_A, 'oned_file',
+                    oned_concat, 'in_files')
+
+        oned_matrix_concat = pe.Node(
+            Function(input_names=['in_files'],
+                        output_names=['out_file'],
+                        function=oned_text_concat,
+                        imports=concat_imports),
+            name=f'oned_matrix_concat_{pipe_num}')
+
+        wf.connect(func_motion_correct_A, 'oned_matrix_save',
+                    oned_matrix_concat, 'in_files')
+
+        out_md1d = pe.Node(
+            interface=util.IdentityInterface(fields=['out_file']),
+            name=f'out_md1d_{pipe_num}')
+
+        wf.connect(md1d_concat, 'out_file',
+                    out_md1d, 'out_file')
+
+        out_oned = pe.Node(
+            interface=util.IdentityInterface(fields=['out_file']),
+            name=f'out_oned_{pipe_num}')
+
+        wf.connect(oned_concat, 'out_file',
+                    out_oned, 'out_file')
+
+        out_oned_matrix = pe.Node(
+            interface=util.IdentityInterface(fields=['out_file']),
+            name=f'out_oned_matrix_{pipe_num}')
+
+        wf.connect(oned_matrix_concat, 'out_file',
+                    out_oned_matrix, 'out_file')
+
+    else:
+        out_motion_A = pe.Node(
+            interface=util.IdentityInterface(fields=['out_file']),
+            name=f'out_motion_A_{pipe_num}')
+
+        wf.connect(func_motion_correct_A, 'out_file',
+                    out_motion_A, 'out_file')
+
+        out_md1d = pe.Node(
+            interface=util.IdentityInterface(fields=['out_file']),
+            name=f'out_md1d_{pipe_num}')
+
+        wf.connect(func_motion_correct_A, 'md1d_file',
+                    out_md1d, 'out_file')
+
+        out_oned = pe.Node(
+            interface=util.IdentityInterface(fields=['out_file']),
+            name=f'out_oned_{pipe_num}')
+
+        wf.connect(func_motion_correct_A, 'oned_file',
+                    out_oned, 'out_file')
+
+        out_oned_matrix = pe.Node(
+            interface=util.IdentityInterface(fields=['out_file']),
+            name=f'out_oned_matrix_{pipe_num}')
+
+        wf.connect(func_motion_correct_A, 'oned_matrix_save',
+                    out_oned_matrix, 'out_file')
+
+    outputs = {
+        'desc-preproc_bold': (out_motion_A, 'out_file'),
+        'desc-motion_bold': (out_motion_A, 'out_file'),
+        'max-displacement': (out_md1d, 'out_file'),
+        'movement-parameters': (out_oned, 'out_file'),
+        'coordinate-transformation': (out_oned_matrix, 'out_file')
+    }
+
+    return wf, outputs
+
+
+def motion_correct_mcflirt(wf, cfg, strat_pool, pipe_num):
+    """Calculate motion parameters with MCFLIRT"""
+    func_motion_correct_A = pe.Node(
+        interface=fsl.MCFLIRT(save_mats=True, save_plots=True),
+        name=f'func_motion_correct_mcflirt_{pipe_num}', mem_gb=2.5)
+
+    func_motion_correct_A.inputs.save_mats = True
+    func_motion_correct_A.inputs.save_plots = True
+    func_motion_correct_A.inputs.save_rms = True
+
+    node, out = strat_pool.get_data('desc-preproc_bold')
+    wf.connect(node, out, func_motion_correct_A, 'in_file')
+
+    node, out = strat_pool.get_data('motion-basefile')
+    wf.connect(node, out, func_motion_correct_A, 'ref_file')
+
+    normalize_motion_params = pe.Node(Function(
+        input_names=['in_file'], output_names=['out_file'],
+        function=normalize_motion_parameters),
+                                      name=f'norm_motion_params_{pipe_num}')
+
+    wf.connect(func_motion_correct_A, 'par_file',
+                normalize_motion_params, 'in_file')
+
+    get_rms_abs = pe.Node(Function(input_names=['rms_files'],
+                                    output_names=['abs_file',
+                                                    'rels_file'],
+                                    function=get_mcflirt_rms_abs),
+                            name=f'get_mcflirt_rms_abs_{pipe_num}')
+
+    wf.connect(func_motion_correct_A, 'rms_files',
+                get_rms_abs, 'rms_files')
+
+    outputs = {
+        'desc-preproc_bold': (func_motion_correct_A, 'out_file'),
+        'desc-motion_bold': (func_motion_correct_A, 'out_file'),
+        'max-displacement': (get_rms_abs, 'abs_file'),
+        'rels-displacement': (get_rms_abs, 'rels_file'),
+        'movement-parameters': (normalize_motion_params, 'out_file'),
+        'coordinate-transformation': (func_motion_correct_A, 'mat_file')
+    }
+
+    return wf, outputs
+
+
+motion_correct = {'3dvolreg': motion_correct_3dvolreg,
+                  'mcflirt': motion_correct_mcflirt}
+
+
+def motion_correct_connections(wf, cfg, strat_pool, pipe_num, opt):
+    """Check opt for valid option, then connect that option.
+
+    Parameters
+    ----------
+    passed through from node block
+
+    Returns
+    -------
+    wf : pe.Workflow
+
+    outputs : dict
+    """
+    if opt not in motion_correct_options:
+        raise KeyError("\n\n[!] Error: The 'tool' parameter of the "
+                       "'motion_correction' workflow must be one of "
+                       f"{str(motion_correct_options).strip('[{()}]')}"
+                       f".\n\nTool input: {opt}\n\n")
+    return motion_correct[opt](wf, cfg, strat_pool, pipe_num)
+
+
+def normalize_motion_parameters(in_file):
+    """Convert FSL mcflirt motion parameters to AFNI space"""
+    import os
+    import numpy as np
+
+    motion_params = np.genfromtxt(in_file).T
+    motion_params = np.vstack((motion_params[2, :] * 180 / np.pi,
+                               motion_params[0, :] * 180 / np.pi,
+                               -motion_params[1, :] * 180 / np.pi,
+                               motion_params[5, :],
+                               motion_params[3, :],
+                               -motion_params[4, :]))
+    motion_params = np.transpose(motion_params)
+
+    out_file = os.path.join(os.getcwd(), 'motion_params.1D')
+    np.savetxt(out_file, motion_params)
+
+    return out_file
